Double properties changed to decimal in Portfolio.API

+++ Portfolio.API/Application/DTOs/AssetDto.cs
@@ -6,6 +6,6 @@ public class AssetDto : BaseDto
 {
     public WalletDto Wallet { get; set; } = new WalletDto();
     public string Symbol { get; set; } = string.Empty;
-    public double AverageBuyPrice { get; set; }
-    public double Quantity { get; set; }
+    public decimal AverageBuyPrice { get; set; }
+    public decimal Quantity { get; set; }
 }
